Certificates_send/requirements.txt
Pillow
sendgrid
pandas

